requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
splitio-client==8.3.1
python-dotenv==1.0.0